requirements.txt
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fonttools==4.33.3
h11==0.13.0
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.1
jupyter-core==4.10.0
kiwisolver==1.4.2
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
murmurhash==1.0.7
nest-asyncio==1.5.5
nltk==3.7
numpy==1.22.3
openpyxl==3.0.10
outcome==1.1.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
preshed==3.0.6
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.8.2
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==22.3.0
regex==2022.6.2
requests==2.27.1
scipy==1.8.1
seaborn==0.11.2
selenium==4.1.5
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.2.0
thinc==8.0.17
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
trio==0.20.0
trio-websocket==0.9.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
wordcloud==1.8.1
wsproto==1.1.0
